requirements.txt
# Flask core
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-WTF>=1.2.1

# Database
SQLAlchemy>=2.0.36

# Data processing and export
pandas>=2.2.0

# Environment configuration
python-dotenv==1.0.0

# Forms and validation
WTForms==3.1.1

# Optimization
PuLP==2.8.0

# Desktop app wrapper
flaskwebgui==1.0.6

# Build / packaging
pyinstaller>=6.10.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
